requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
annoy==1.17.3
anyio==4.13.0
appdirs==1.4.4
asttokens==3.0.1
attrs==26.1.0
bleach==6.3.0
boto3==1.42.75
botocore==1.42.75
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
datasets==4.8.4
decorator==5.2.1
dill==0.4.1
docopt==0.6.2
executing==2.2.1
open-clip-torch==3.3.0
fastapi==0.135.2
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
ipyplot==1.1.2
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
kaggle==2.0.0
kagglesdk==0.1.16
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
numpy==2.4.3
packaging==26.0
pandas==3.0.1
parso==0.8.6
pgvector==0.4.2
pillow==12.1.1
pipreqs==0.4.13
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
python-slugify==8.0.4
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
s3transfer==0.16.0
safetensors==0.7.0
setuptools==81.0.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
text-unidecode==1.3
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
validators==0.35.0
wcwidth==0.6.0
webencodings==0.5.1
xxhash==3.6.0
yarg==0.1.10
yarl==1.23.0
